requirements.txt
blessings==1.7
bpython==0.18
certifi==2019.9.11
chardet==3.0.4
curtsies==0.3.0
Django==2.2.5
djangorestframework==3.10.3
greenlet==0.4.15
idna==2.8
psycopg2==2.8.3
Pygments==2.4.2
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typing==3.7.4.1
urllib3==1.25.6
wcwidth==0.1.7
